A utility to collect time samples of pathogen levels, plus...
moved some policy-specific tables of constants to $(MODELDIR).  Added
code to generate a table of counts of discharges which are not
transfers.  Added support for setting the random number seed.  Added
trackedFacilities, a list of facilities for which time series and misc.
other info is to be collected.  Made $(MODELDIR) an absolute path.