Added a registry class that can be attached to a facility to allow for each facility to keep track of their CRE Positive patients (based on an urgent request from Sarah and Bruce).
Updated GenericDiagnosticPolicy to place patients on facility registries based on probabilities by facility category and a probability of a person is transferred, that the information will be communicated to the new facility.

If a patient enters a facility, and is on the registry, then they are put on contact precautions throughout the remainder of the simulation.