Mods to support facilitybase.PatientCumStats
removed a diagnostic print from hospital.py.  Ignored freeze-dried
patients when calculating pathogen exposure in community (which we never
did in the community in the past).  Creation of Community agents is
split from their freeze-drying to allow pathogens to get added.  This
freezes pathogen numbers in the community into cache.  It also required
Facilities to get a finalizeBuild() method.  pyrheabase.py gets a bugfix
to a problem with partitioning of wards by patient type; rejected
assignments were causing wards to be dropped from the pool.